+++ src/Leefrost.Featurebase/Leefrost.Featurebase.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <Folder Include="Clients\Cloud\" />
+    <Folder Include="Query\Pql\" />
   </ItemGroup>
 
   <ItemGroup>
