+++ src/Leefrost.Featurebase/Pql/Selection/ConstRow.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text;
 
-namespace Leefrost.Featurebase.Query.Pql;
+namespace Leefrost.Featurebase.Pql.Selection;
 public class ConstRow : RowQuery
 {
     private readonly List<string> _columns;
