+++ src/Leefrost.Featurebase/Pql/Selection/Difference.cs
@@ -1,29 +1,29 @@
 ﻿using System.Text;
 
-namespace Leefrost.Featurebase.Query.Pql;
-public class Union : RowQuery
+namespace Leefrost.Featurebase.Pql.Selection;
+public class Difference : RowQuery
 {
     private readonly List<RowQuery> _rows = [];
 
-    public Union(RowQuery row1, RowQuery row2)
+    public Difference(RowQuery row1, RowQuery row2)
     {
         _rows.Add(row1);
         _rows.Add(row2);
     }
 
-    public Union(IEnumerable<RowQuery> rows)
+    public Difference(IEnumerable<RowQuery> rows)
     {
         var queries = rows.ToList();
-        if (queries.Count == 0)
-            throw new ArgumentException("Union must have at least one row argument");
+        if (queries.Count < 2)
+            throw new ArgumentException("Difference must have at least 2 rows to compare");
 
         _rows = queries;
     }
-    
+
     public override string Build()
     {
         var builder = new StringBuilder();
-        builder.Append("Union(");
+        builder.Append("Difference(");
         builder.Append(string.Join(',', _rows.Select(row => row.Build())));
         builder.Append(')');
