+++ src/Leefrost.Featurebase/Pql/Selection/Intersect.cs
@@ -1,29 +1,29 @@
 ﻿using System.Text;
 
-namespace Leefrost.Featurebase.Query.Pql;
-public class Difference : RowQuery
+namespace Leefrost.Featurebase.Pql.Selection;
+public class Intersect : RowQuery
 {
     private readonly List<RowQuery> _rows = [];
 
-    public Difference(RowQuery row1, RowQuery row2)
+    public Intersect(RowQuery row1, RowQuery row2)
     {
         _rows.Add(row1);
         _rows.Add(row2);
     }
 
-    public Difference(IEnumerable<RowQuery> rows)
+    public Intersect(IEnumerable<RowQuery> rows)
     {
         var queries = rows.ToList();
-        if (queries.Count < 2)
-            throw new ArgumentException("Difference must have at least 2 rows to compare");
+        if (queries.Count == 0)
+            throw new ArgumentException("Intersect must have at least one row argument");
 
         _rows = queries;
     }
-    
+
     public override string Build()
     {
         var builder = new StringBuilder();
-        builder.Append("Difference(");
+        builder.Append("Intersect(");
         builder.Append(string.Join(',', _rows.Select(row => row.Build())));
         builder.Append(')');
