+++ src/Leefrost.Featurebase/Pql/Selection/Row.cs
@@ -2,7 +2,7 @@
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace Leefrost.Featurebase.Query.Pql;
+namespace Leefrost.Featurebase.Pql.Selection;
 
 public class Row : RowQuery
 {
