+++ src/Leefrost.Featurebase/Pql/Selection/Union.cs
@@ -1,29 +1,29 @@
 ﻿using System.Text;
 
-namespace Leefrost.Featurebase.Query.Pql;
-public class Xor : RowQuery
+namespace Leefrost.Featurebase.Pql.Selection;
+public class Union : RowQuery
 {
     private readonly List<RowQuery> _rows = [];
 
-    public Xor(RowQuery row1, RowQuery row2)
+    public Union(RowQuery row1, RowQuery row2)
     {
         _rows.Add(row1);
         _rows.Add(row2);
     }
 
-    public Xor(IEnumerable<RowQuery> rows)
+    public Union(IEnumerable<RowQuery> rows)
     {
         var queries = rows.ToList();
-        if (queries.Count < 2)
-            throw new ArgumentException("Xor must have at least 2 row arguments");
+        if (queries.Count == 0)
+            throw new ArgumentException("Union must have at least one row argument");
 
         _rows = queries;
     }
-    
+
     public override string Build()
     {
         var builder = new StringBuilder();
-        builder.Append("Xor(");
+        builder.Append("Union(");
         builder.Append(string.Join(',', _rows.Select(row => row.Build())));
         builder.Append(')');
