+++ src/Leefrost.Featurebase/Pql/Selection/Xor.cs
@@ -1,29 +1,29 @@
 ﻿using System.Text;
 
-namespace Leefrost.Featurebase.Query.Pql;
-public class Intersect : RowQuery
+namespace Leefrost.Featurebase.Pql.Selection;
+public class Xor : RowQuery
 {
     private readonly List<RowQuery> _rows = [];
 
-    public Intersect(RowQuery row1, RowQuery row2)
+    public Xor(RowQuery row1, RowQuery row2)
     {
         _rows.Add(row1);
         _rows.Add(row2);
     }
 
-    public Intersect(IEnumerable<RowQuery> rows)
+    public Xor(IEnumerable<RowQuery> rows)
     {
         var queries = rows.ToList();
-        if (queries.Count == 0)
-            throw new ArgumentException("Intersect must have at least one row argument");
+        if (queries.Count < 2)
+            throw new ArgumentException("Xor must have at least 2 row arguments");
 
         _rows = queries;
     }
-    
+
     public override string Build()
     {
         var builder = new StringBuilder();
-        builder.Append("Intersect(");
+        builder.Append("Xor(");
         builder.Append(string.Join(',', _rows.Select(row => row.Build())));
         builder.Append(')');
